Correção de diversos métodos para se adequar com as APIs da Pagcerto

+++ pagcertosdk/src/main/java/br/com/paggcerto/pagcertosdk/model/payments/request/TransactionsToAnticipate.kt
@@ -7,5 +7,5 @@ import java.io.Serializable
 class TransactionsToAnticipate: Serializable {
     @SerializedName("cardTransactions")
     @Expose
-    var cardTransactions: List<Long> = emptyList()
+    var cardTransactions: List<Int> = emptyList()
 }
